# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SpendWise AI API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Transaction {
  id: number;
  merchant: string;
  amount: number;
  category: string;
  /** normal or impulse */
  status: string;
  isImpulse: boolean;
  nudgeTriggered: boolean;
  /** @nullable */
  investmentSuggested: number | null;
  /** @nullable */
  notes: string | null;
  createdAt: string;
}

export interface CreateTransactionBody {
  merchant: string;
  amount: number;
  category: string;
  notes?: string;
}

export interface Category {
  id: number;
  name: string;
  color: string;
  icon: string;
  blockSpending: boolean;
  allowInvestNudges: boolean;
  createdAt: string;
}

export interface CreateCategoryBody {
  name: string;
  color?: string;
  icon?: string;
  blockSpending?: boolean;
  allowInvestNudges?: boolean;
}

export interface UpdateCategoryBody {
  name?: string;
  color?: string;
  icon?: string;
  blockSpending?: boolean;
  allowInvestNudges?: boolean;
}

export interface Budget {
  id: number;
  category: string;
  limitAmount: number;
  spentAmount: number;
  /** monthly or weekly */
  period: string;
  /** soft or hard */
  budgetType: string;
  isActive: boolean;
  createdAt: string;
}

export interface CreateBudgetBody {
  category: string;
  limitAmount: number;
  period: string;
  budgetType: string;
}

export interface UpdateBudgetBody {
  category?: string;
  limitAmount?: number;
  period?: string;
  budgetType?: string;
  isActive?: boolean;
}

export interface NudgeRule {
  id: number;
  name: string;
  thresholdAmount: number;
  /** fixed or percentage */
  investmentType: string;
  investmentValue: number;
  /** gentle or aggressive */
  nudgeIntensity: string;
  isActive: boolean;
  createdAt: string;
}

export interface CreateNudgeRuleBody {
  name: string;
  thresholdAmount: number;
  investmentType: string;
  investmentValue: number;
  nudgeIntensity: string;
}

export interface UpdateNudgeRuleBody {
  name?: string;
  thresholdAmount?: number;
  investmentType?: string;
  investmentValue?: number;
  nudgeIntensity?: string;
  isActive?: boolean;
}

export interface Nudge {
  id: number;
  transactionId: number;
  /** @nullable */
  nudgeRuleId: number | null;
  message: string;
  suggestedInvestment: number;
  /** pending, accepted, dismissed */
  status: string;
  createdAt: string;
}

export interface Investment {
  id: number;
  amount: number;
  /** manual or auto */
  source: string;
  category: string;
  /** @nullable */
  transactionId: number | null;
  /** @nullable */
  nudgeId: number | null;
  /** @nullable */
  notes: string | null;
  createdAt: string;
}

export interface InvestmentSettings {
  id: number;
  overspendPercent: number;
  weeklyCap: number;
  autoInvestEnabled: boolean;
  updatedAt: string;
}

export interface UpdateInvestmentSettingsBody {
  overspendPercent?: number;
  weeklyCap?: number;
  autoInvestEnabled?: boolean;
}

export interface Gamification {
  id: number;
  totalPoints: number;
  level: string;
  streak: number;
  impulsesAvoided: number;
  budgetsRespected: number;
  investmentsMade: number;
  streakBonuses: number;
  updatedAt: string;
}

export interface LeaderboardEntry {
  rank: number;
  name: string;
  points: number;
  level: string;
  streak: number;
}

export interface DashboardSummary {
  todaySpend: number;
  totalInvested: number;
  impulseScore: number;
  activeAlerts: number;
  weeklySpend: number;
  monthlySpend: number;
  pendingNudgesCount: number;
  totalTransactions: number;
}

export interface SpendingTrendPoint {
  date: string;
  totalSpend: number;
  impulseSpend: number;
  normalSpend: number;
}

export interface CategoryBreakdownItem {
  category: string;
  amount: number;
  count: number;
  color: string;
}

export interface AdminStats {
  totalNudgesFired: number;
  investmentsTriggered: number;
  totalTransactions: number;
  totalUsers: number;
  avgImpulseScore: number;
  totalAmountInvested: number;
  nudgeAcceptanceRate: number;
}

export type ListTransactionsParams = {
  category?: string;
  status?: string;
  limit?: number;
};