# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SpendWise AI API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List all categories
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [categories]
      summary: Create a category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoryBody"
      responses:
        "201":
          description: Created category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /categories/{id}:
    get:
      operationId: getCategory
      tags: [categories]
      summary: Get a category by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateCategory
      tags: [categories]
      summary: Update a category
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCategoryBody"
      responses:
        "200":
          description: Updated category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
    delete:
      operationId: deleteCategory
      tags: [categories]
      summary: Delete a category
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Budgets
  /budgets:
    get:
      operationId: listBudgets
      tags: [budgets]
      summary: List all budgets
      responses:
        "200":
          description: List of budgets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Budget"
    post:
      operationId: createBudget
      tags: [budgets]
      summary: Create a budget
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBudgetBody"
      responses:
        "201":
          description: Created budget
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Budget"
  /budgets/{id}:
    patch:
      operationId: updateBudget
      tags: [budgets]
      summary: Update a budget
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBudgetBody"
      responses:
        "200":
          description: Updated budget
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Budget"
    delete:
      operationId: deleteBudget
      tags: [budgets]
      summary: Delete a budget
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Nudge Rules
  /nudge-rules:
    get:
      operationId: listNudgeRules
      tags: [nudges]
      summary: List all nudge rules
      responses:
        "200":
          description: List of nudge rules
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NudgeRule"
    post:
      operationId: createNudgeRule
      tags: [nudges]
      summary: Create a nudge rule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateNudgeRuleBody"
      responses:
        "201":
          description: Created nudge rule
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NudgeRule"
  /nudge-rules/{id}:
    patch:
      operationId: updateNudgeRule
      tags: [nudges]
      summary: Update a nudge rule
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateNudgeRuleBody"
      responses:
        "200":
          description: Updated nudge rule
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NudgeRule"
    delete:
      operationId: deleteNudgeRule
      tags: [nudges]
      summary: Delete a nudge rule
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /nudges/{id}/dismiss:
    post:
      operationId: dismissNudge
      tags: [nudges]
      summary: Dismiss a nudge
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Dismissed nudge
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Nudge"
  /nudges/{id}/accept:
    post:
      operationId: acceptNudge
      tags: [nudges]
      summary: Accept a nudge (invest)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Accepted nudge
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Nudge"

  # Investments
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Category:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        color:
          type: string
        icon:
          type: string
        blockSpending:
          type: boolean
        allowInvestNudges:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - color
        - icon
        - blockSpending
        - allowInvestNudges
        - createdAt
    CreateCategoryBody:
      type: object
      properties:
        name:
          type: string
        color:
          type: string
        icon:
          type: string
        blockSpending:
          type: boolean
        allowInvestNudges:
          type: boolean
      required:
        - name
    UpdateCategoryBody:
      type: object
      properties:
        name:
          type: string
        color:
          type: string
        icon:
          type: string
        blockSpending:
          type: boolean
        allowInvestNudges:
          type: boolean
    Budget:
      type: object
      properties:
        id:
          type: integer
        category:
          type: string
        limitAmount:
          type: number
        spentAmount:
          type: number
        period:
          type: string
          description: "monthly or weekly"
        budgetType:
          type: string
          description: "soft or hard"
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - category
        - limitAmount
        - spentAmount
        - period
        - budgetType
        - isActive
        - createdAt
    CreateBudgetBody:
      type: object
      properties:
        category:
          type: string
        limitAmount:
          type: number
        period:
          type: string
        budgetType:
          type: string
      required:
        - category
        - limitAmount
        - period
        - budgetType
    UpdateBudgetBody:
      type: object
      properties:
        category:
          type: string
        limitAmount:
          type: number
        period:
          type: string
        budgetType:
          type: string
        isActive:
          type: boolean
    NudgeRule:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        thresholdAmount:
          type: number
        investmentType:
          type: string
          description: "fixed or percentage"
        investmentValue:
          type: number
        nudgeIntensity:
          type: string
          description: "gentle or aggressive"
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - thresholdAmount
        - investmentType
        - investmentValue
        - nudgeIntensity
        - isActive
        - createdAt
    CreateNudgeRuleBody:
      type: object
      properties:
        name:
          type: string
        thresholdAmount:
          type: number
        investmentType:
          type: string
        investmentValue:
          type: number
        nudgeIntensity:
          type: string
      required:
        - name
        - thresholdAmount
        - investmentType
        - investmentValue
        - nudgeIntensity
    UpdateNudgeRuleBody:
      type: object
      properties:
        name:
          type: string
        thresholdAmount:
          type: number
        investmentType:
          type: string
        investmentValue:
          type: number
        nudgeIntensity:
          type: string
        isActive:
          type: boolean
    Nudge:
      type: object
      properties:
        id:
          type: integer
        transactionId:
          type: integer
        nudgeRuleId:
          type: ["integer", "null"]
        message:
          type: string
        suggestedInvestment:
          type: number
        status:
          type: string
          description: "pending, accepted, dismissed"
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - transactionId
        - nudgeRuleId
        - message
        - suggestedInvestment
        - status
        - createdAt
...